refactor: split minigame data
